esrgan-backend/requirements.txt
# Real-ESRGAN and its dependencies
torch==1.10.1
torchvision==0.11.2
opencv-python
numpy
Pillow
realesrgan
basicsr>=1.4.2
facexlib>=0.2.5
gfpgan>=1.3.5

# Flask server
flask

# Optional for progress/debugging
tqdm